Fixed Decryption
I fixed the decryption feature, it will now use output.txt as the filname to input and create a new file named decrypted.txt to contain the decryption. In case the user tries to decrypt first, it will check if output exists and if it doesnt, will prompt the user with an error message.